refactor(swag): split 937-line swag.py into focused modules (2hx)
Move all _handle_*_action functions out of swag.py into handlers/:
- handlers/configs.py — list, create, view, edit, remove, update (+ health check helpers)
- handlers/logs.py — logs handler
- handlers/backups.py — backups handler
- handlers/health.py — health_check handler

swag.py now contains only register_tools(), the tool signature, and dispatch.
