+++ swag_mcp/tools/handlers/__init__.py
@@ -1,0 +1,1 @@
+"""Handler modules for SWAG tool actions."""
